event_hub_scaling_task/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.8.5
azure-functions==1.18.0
azure-identity==1.15.0
azure-mgmt-eventhub==11.0.0
jsonschema==4.21.1
